+++ 2_coroutine/scheduler.h
@@ -38,19 +38,19 @@ struct tbuffer
 // http://blog.olkie.com/2013/11/05/online-c-function-prototype-header-generator-tool/
 void sched_clear(struct sbuffer **st, struct sbuffer **en);
 void sched_count(struct sbuffer **st, struct sbuffer **en);
-void sched_rpop(struct sbuffer **st, struct sbuffer **en,  jmp_buf *flag, int *key, int *id);
-void sched_insert(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, int key, int id );
-void sched_rpush(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, int key, int id );
-void sched_lpush(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, int key, int id );
+void sched_rpop(struct sbuffer **st, struct sbuffer **en,  jmp_buf *flag, long *key, int *id);
+void sched_insert(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, long key, int id );
+void sched_rpush(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, long key, int id );
+void sched_lpush(struct sbuffer **st, struct sbuffer **en,  jmp_buf flag, long key, int id );
 void sched_init(SCHED* self, int finish);
 SCHED* sched_create(int finish);
-void sched_yield(SCHED* self, jmp_buf flag, int then);
-void spawn(SCHED* self, void (func_ptr()), void *typex, int then);
+void sched_yield(SCHED* self, jmp_buf flag, long then);
+void spawn(SCHED* self, void (func_ptr()), void *typex, long then);
 void sched_run(SCHED* self);
 void sched_sweep(SCHED* self);
 void task_lpush(struct tbuffer **st, struct tbuffer **en,  void* typex, void (*func_ptr)());
 void task_rpop(struct tbuffer **st, struct tbuffer **en,  void **typex, void (**func_ptr)());
-void waitfor(SCHED* self, int n);
-void waituntil(SCHED* self, int n);
+void waitfor(SCHED* self, long n);
+void waituntil(SCHED* self, long n);
 
 #endif // SCHEDULER_H
